replaced didChangeDependencies to root account widget
