+++ lib/ui/widgets/account/account_widget.dart
@@ -1,8 +1,12 @@
 // Flutter imports:
 import 'package:flutter/material.dart';
 
+// Package imports:
+import 'package:flutter_bloc/flutter_bloc.dart';
+
 // Project imports:
 import 'package:comics_db_app/ui/components/custom_appbar_widget.dart';
+import 'package:comics_db_app/ui/widgets/account/account_details_cubit.dart';
 import 'package:comics_db_app/ui/widgets/account/components/body_account_details_widget.dart';
 
 // Package imports:
@@ -16,6 +20,23 @@ class AccountWidget extends StatefulWidget {
 }
 
 class _AccountWidgetState extends State<AccountWidget> {
+  late Future<String> lazyValue;
+
+  @override
+  void didChangeDependencies() {
+    super.didChangeDependencies();
+    Future<String> loadingDelay() {
+      Duration duration = const Duration(seconds: 2);
+
+      return Future.delayed(duration, () => 'It took ${duration.inSeconds}');
+    }
+
+    lazyValue = loadingDelay();
+
+    final locale = Localizations.localeOf(context);
+    context.read<AccountDetailsCubit>().setupAccountDetails(context, locale.languageCode);
+  }
+
   @override
   Widget build(BuildContext context) {
     return const Scaffold(
